pet-race-job/requirements.txt
nose
sphinx
cassandra-driver
numpy